requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\CNNTL.py: 12
# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\main.py: 13
Pillow == 8.3.1

# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\CNNTL.py: 4,138,139,140
# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\main.py: 158
keras == 2.6.0

# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\CNNTL.py: 3
# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\main.py: 157
matplotlib == 3.4.3

# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\CNNTL.py: 2,9
# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\main.py: 10,156
numpy == 1.19.5

# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\CNNTL.py: 11
# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\main.py: 12
opencv_python == 4.5.3.56

# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\CNNTL.py: 153
pandas == 1.3.2

# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\CNNTL.py: 102
# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\main.py: 203
scikit_learn == 1.0

# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\main.py: 239,240,241
# tensorflow == 2.6.0

# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\CNNTL.py: 15
# C:\Users\Sarah Allam\Desktop\Bachelor Project\Code\Model 1\main.py: 16
tensorflow_datasets == 4.4.0
